begin setting up strictmode handler
